静态对象配置文件、加载、管理

+++ Learn_PaoPaoTang/Classes/GameObject.h
@@ -36,7 +36,7 @@ public:
 	}
 
 public:
-	virtual int getDepth() { return -mRenderObj.getPosition().y; }
+	virtual float getDepth() { return -mRenderObj.getPosition().y; }
 	virtual void load(const char* szName) = 0;
 	virtual void update(float dt)
 	{
